requirements.txt
pytest==7.1.2
mlflow==1.27.0
pandas==1.4.3
sklearn==0.0
scikit-learn==1.1.1
matplotlib==3.7.2
pytorch==2.0.1